data set upload
added data files, driveTest contains the longest continuous set of data, the last 5 seconds or so should probably be discarded the motor controller overheated. The other 2 sets, testDrive2 and testDrive3 are similar, one includes more left turns while the other is more right turns. There is significant noise on the x,y position when crossing x = 0. The recorded files in their respective directories are .csv files. The data is recorded using ROS

jetRacer_Controller is the recorded inputs to the racer
jetRacer_sensor-imu_data is the recorded accelerometer and gyro data in x,y, and z. Also recorded is magnetometer but this is not useful because it is seriously corrupted by the noise of the motor
jetRacer_sensor-front/backTOF_data is the recorded front facing and back facing range finder respectively, unsure if this is useful
jetRacer_sensor-wheel_data is the recorded front wheel odometery, measurements are in degrees from 0-360. when the angle exceeds 360 it jumps back to 0
jetRacer_sensor-opticalFlow_data is the recorded optical flow data from a modified optical mouse sensor which in theory is capable of measuring deltaX and deltaY measurements, not sure it's useful data either